requirements.txt
fastapi
uvicorn
supabase-py
python-jose
passlib
sqlalchemy